previewFile apiを修正
